django-testing/django_testing/requirements.txt
asgiref==3.5.2
attrs==22.1.0
coverage==6.5.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.1
iniconfig==1.1.1
model-bakery==1.9.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
